requirements.txt
# Core
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.25.0
Pillow>=10.0.0
joblib>=1.3.0
pytest>=8.2.0

# Hugging Face
transformers>=4.40.0
huggingface-hub>=0.18.0

# FastAPI and related
fastapi>=0.102.0
pydantic>=2.5.0
python-multipart>=0.0.6
uvicorn[standard]>=0.23.0

# Utilities
python-dotenv>=1.0.0
httpx